nvidia-utils: Enable nvidia services as default in .install
The post_install will enable the services as default on first installation. This is an requirement for having sleep working on NVIDIa GPUs, specially on wayland.

post_upgrade will only enable the services once on upgrade for the upcoming version. This is intended that the user only gets in enabled once, to reflect our default behavior. If the user wants to disable them again, they want be forced to be installed.

Besides that we are removing some old comments, which are not revelant anymore.

Signed-off-by: Peter Jung <ptr1337@archlinux.org>