nvidia-utils: 565.57.01
NVIDIA has recently released a new beta driver that includes several functional and security fixes.

The current 560 Driver is vulnerable to CVE-2024-0126. While NVIDIA has provided fixes for the 535, 550, and 565 branches, they have not issued a fix for the stable 560 branch. Additionally, NVIDIA does not plan to release a 560 update for Linux that addresses this CVE.

As users of the stable 560 branch remain vulnerable, we propose moving the 565 driver into testing, with the intention of transitioning it to stable after a testing period. The final release of the 565 branch is expected in approximately 3-4 weeks.

Known Issues with the 565 Driver:
- Displays connected via DVI may not function on certain graphics cards.

The vulnerability has a severity rating of 8.2 (High). NVIDIA describes it as follows: "NVIDIA GPU Display Driver for Windows and Linux contains a vulnerability that could allow a privileged attacker to escalate permissions. A successful exploit of this vulnerability might lead to code execution, denial of service, escalation of privileges, information disclosure, and data tampering."

Links:
https://nvidia.custhelp.com/app/answers/detail/a_id/5586

Signed-off-by: Peter Jung <ptr1337@archlinux.org>